Added Serial support for K22F and fixed overal serial
First part is pretty obvious, added some pin defines, and some ifdefs in
serial_api.c to make it compile (they all use KSDK macros, so goes
automatically, and not that many ifdefs).

Second part is that in the new version of the hal files you apparantly
got to manually enable the uart transmitter/receiver. This wasn't in
there yet, so it did very little.

Now it works fine on a K22F (well minus the part where the first char in
a program isn't transmitted, but thats alot better than nothing being
transmitted)