[HAL][KSDK] Use PIT timer for ticker/timeout
Previously in the KLXX devices the LPTMR was used for ticker/timeout,
since they only have 2 PIT channels which are required for regular
timer. The KSDK targets have 4 PIT channels, which allows for two 32-bit
timers running at 1 MHz.

Since PIT timers are 32-bit, and LPTMR 16-bit, this both simplifies it
and makes it more efficient. It does cost the remaining 2 PIT channels,
but on the other hand leaves the LPTMR free.